refactor(web): extract shared useGenerationStudio hook + preset components (sc-1650)
ImageStudio and VideoStudio carried ~150-200 lines of near-identical preset and
result-tracking logic that drifted whenever a bug fix landed in only one. Pull
the shared parts into apps/web/src/screens/generationStudio.jsx:

- useGenerationStudio(): preset selection + availablePresets/selectedPreset/
  presetPromptParts/presetLoraDetails/presetValidationResult, the "snap an
  explicit preset back into range" effect, the catalog-driven model- and
  character-reset effects, and the completed-job result-visibility machinery
  (resultVisible/completedAnchorMs/completedWaitExpired + fallback timer +
  localJobs). VideoStudio's localJobs is now memoized too (it wasn't before).
- onPromptKeyDown: shared Cmd/Ctrl+Enter submit handler.
- PresetGuidanceStrip / PresetValidationWarnings: the verbatim-duplicated
  guidance strip and missing/incompatible warning JSX (no-preset hint text
  parameterized).

Each studio keeps its divergent pieces: preset-default field application
(image count/resolution vs video duration/fps/quality), launch-request
handling, submit payloads, and Image's extra selected-LoRA validation warning.
Behavior-preserving: the moved effects are dependency-driven with deferred
setState, so declaration-order changes don't alter outcomes.

Net -174 lines across the two studios. 107 web tests pass; vite build clean.
