+++ apps/web/src/screens/VideoStudio.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useRef, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { AssetPickerField } from "../components/AssetPicker.jsx";
 import { AssetCard } from "../components/assetPanels.jsx";
 import { AssetMedia, assetCanRenderAsVideo } from "../components/assetMedia.jsx";
@@ -40,16 +40,16 @@ import {
   clearPresetDefault,
   loraLooksLikeIcLora,
   noPresetId,
-  presetLoraDetails as buildPresetLoraDetails,
-  presetMatchesModel,
-  presetMatchesWorkflow,
-  presetPromptParts as buildPresetPromptParts,
-  presetValidation,
   rememberPresetDefault,
 } from "../presetUtils.js";
+import {
+  onPromptKeyDown,
+  PresetGuidanceStrip,
+  PresetValidationWarnings,
+  useGenerationStudio,
+} from "./generationStudio.jsx";
 import { ReplacePersonPanel, findReplacementModel } from "./ReplacePersonPanel.jsx";
 
-const completedResultFallbackMs = 30000;
 const ltxVideoModelId = "ltx_2_3";
 const ltxIcLoraRequiredModes = new Set(["extend_clip", "video_bridge"]);
 
@@ -95,7 +95,6 @@ export function VideoStudio({
   const [precision, setPrecision] = useState("fp8");
   const [advancedOpen, setAdvancedOpen] = useState(false);
   const [model, setModel] = useState(videoModels[0]?.id ?? ltxVideoModelId);
-  const [selectedPresetId, setSelectedPresetId] = useState(null);
   const selectedModel = videoModels.find((item) => item.id === model) ?? videoModels[0];
   const [duration, setDuration] = useState(selectedModel?.defaults?.duration ?? 6);
   const [resolution, setResolution] = useState(selectedModel?.defaults?.resolution ?? "768x512");
@@ -114,7 +113,6 @@ export function VideoStudio({
   const [comparisonMode, setComparisonMode] = useState("side_by_side");
   const [abSide, setAbSide] = useState("replacement");
   const [submitting, setSubmitting] = useState(false);
-  const [resultFallbackTick, setResultFallbackTick] = useState(0);
   const previewVideoRef = useRef(null);
   const [previewPlaying, setPreviewPlaying] = useState(false);
   const [previewTime, setPreviewTime] = useState(0);
@@ -123,29 +121,33 @@ export function VideoStudio({
   const capabilities = selectedModel?.capabilities ?? [];
   const supportsMode = capabilities.includes(mode);
   const implementedMode = ["image_to_video", "text_to_video", "first_last_frame", "extend_clip", "replace_person"].includes(mode);
-  const availablePresets = useMemo(() => {
-    return presets.filter((preset) => presetMatchesWorkflow(preset, mode) && presetMatchesModel(preset, selectedModel));
-  }, [mode, presets, selectedModel?.id]);
-  const selectedPreset =
-    selectedPresetId === noPresetId
-      ? null
-      : selectedPresetId
-        ? availablePresets.find((preset) => preset.id === selectedPresetId) ?? null
-        : availablePresets[0] ?? null;
-  const presetPromptParts = buildPresetPromptParts(selectedPreset);
-  const presetLoraDetails = buildPresetLoraDetails(selectedPreset, loras);
-  const presetValidationResult = useMemo(
-    () => presetValidation(selectedPreset, loras, selectedModel),
-    [selectedPreset, loras, selectedModel],
-  );
+  const {
+    availablePresets,
+    selectedPreset,
+    setSelectedPresetId,
+    presetPromptParts,
+    presetLoraDetails,
+    presetValidationResult,
+    localJobs,
+  } = useGenerationStudio({
+    mode,
+    presets,
+    selectedModel,
+    loras,
+    models: videoModels,
+    model,
+    setModel,
+    fallbackModelId: ltxVideoModelId,
+    characters,
+    characterId,
+    setCharacterId,
+    setCharacterLookId,
+    assets,
+    latestAssets,
+    trackedLocalJobs,
+  });
   const requiresLtxIcLora = selectedModel?.id === ltxVideoModelId && ltxIcLoraRequiredModes.has(mode);
   const hasLtxIcLora = presetLoraDetails.some((lora) => !lora.missing && loraLooksLikeIcLora(lora));
-
-  useEffect(() => {
-    if (!videoModels.some((item) => item.id === model)) {
-      setModel(videoModels[0]?.id ?? ltxVideoModelId);
-    }
-  }, [videoModels, model]);
 
   useEffect(() => {
     if (selectedAsset?.type === "image" || selectedAsset?.type === "frame") {
@@ -179,13 +181,6 @@ export function VideoStudio({
   }, [launchRequest?.id, selectedAsset?.id, selectedAsset?.type]);
 
   useEffect(() => {
-    if (characterId && !characters.some((character) => character.id === characterId)) {
-      setCharacterId("");
-      setCharacterLookId("");
-    }
-  }, [characters, characterId]);
-
-  useEffect(() => {
     if (!selectedModel) {
       return;
     }
@@ -212,15 +207,6 @@ export function VideoStudio({
       setModel(replacementModel.id);
     }
   }, [mode, supportsMode, videoModels]);
-
-  useEffect(() => {
-    if (!selectedPresetId || selectedPresetId === noPresetId) {
-      return;
-    }
-    if (!selectedPreset) {
-      setSelectedPresetId(availablePresets[0]?.id ?? noPresetId);
-    }
-  }, [availablePresets, selectedPresetId, selectedPreset]);
 
   useEffect(() => {
     if (!selectedPreset) {
@@ -345,51 +331,6 @@ export function VideoStudio({
     full_person_replace_outfit: "Full Person, Replace Outfit",
   };
 
-  function resultVisible(job) {
-    if (job.result?.generationSetId) {
-      return latestAssets.some((asset) => asset.generationSetId === job.result.generationSetId);
-    }
-    const assetIds = job.result?.assetIds ?? [];
-    return assetIds.length > 0 && assetIds.every((id) => assets.some((asset) => asset.id === id));
-  }
-
-  function completedAnchorMs(job) {
-    return Date.parse(job.completedAt ?? job.updatedAt ?? "");
-  }
-
-  function completedWaitExpired(job, nowMs = Date.now()) {
-    const anchorMs = completedAnchorMs(job);
-    return Number.isFinite(anchorMs) && nowMs - anchorMs > completedResultFallbackMs;
-  }
-
-  useEffect(() => {
-    const nowMs = Date.now();
-    const pendingCompletedJobs = trackedLocalJobs.filter(
-      (job) =>
-        job.status === "completed" &&
-        Number.isFinite(completedAnchorMs(job)) &&
-        !resultVisible(job) &&
-        !completedWaitExpired(job, nowMs),
-    );
-    if (!pendingCompletedJobs.length) {
-      return undefined;
-    }
-    const nextDelay = Math.min(
-      ...pendingCompletedJobs.map((job) => Math.max(0, completedResultFallbackMs - (nowMs - completedAnchorMs(job)))),
-    );
-    const timer = window.setTimeout(() => setResultFallbackTick((value) => value + 1), nextDelay + 50);
-    return () => window.clearTimeout(timer);
-  }, [assets, latestAssets, trackedLocalJobs, resultFallbackTick]);
-
-  const localJobs = trackedLocalJobs.filter(
-    (job) =>
-      // Canceled runs produce no output, so drop them instead of leaving a
-      // "Canceled" progress card behind.
-      job.status !== "canceled" &&
-      (job.status !== "completed" || (!resultVisible(job) && !completedWaitExpired(job))),
-  );
-  const hasReviewContent = Boolean(localJobs.length || latestAssets.length);
-
   async function submit(event) {
     event.preventDefault();
     if (submitting) {
@@ -459,13 +400,6 @@ export function VideoStudio({
       return;
     }
     video.pause();
-  }
-
-  function onPromptKeyDown(event) {
-    if ((event.metaKey || event.ctrlKey) && event.key === "Enter") {
-      event.preventDefault();
-      event.currentTarget.form?.requestSubmit();
-    }
   }
 
   return (
@@ -715,16 +649,7 @@ export function VideoStudio({
             ) : null}
 
             {blockedMessage ? <p className="inline-warning">{blockedMessage}</p> : null}
-            {presetValidationResult.missing.length ? (
-              <p className="inline-warning">
-                Preset cannot run until LoRA import finishes: {presetValidationResult.missing.join(", ")}. Wait for the Queue or choose another preset.
-              </p>
-            ) : null}
-            {presetValidationResult.incompatible.length ? (
-              <p className="inline-warning">
-                Preset cannot run with {selectedModel?.name ?? "the selected model"} because these LoRAs are incompatible: {presetValidationResult.incompatible.join(", ")}. Choose another preset or model.
-              </p>
-            ) : null}
+            <PresetValidationWarnings presetValidationResult={presetValidationResult} selectedModel={selectedModel} />
           </div>
 
           <div className="video-rail">
@@ -826,23 +751,12 @@ export function VideoStudio({
                 </select>
               </label>
 
-              {selectedPreset ? (
-                <div className="guidance-strip">
-                  <strong>{selectedPreset.ui?.description ?? "Preset defaults active"}</strong>
-                  <span>
-                    {presetPromptParts.length ? `Adds: ${presetPromptParts.join(", ")}` : "No prompt fragments"}
-                    {presetLoraDetails.length
-                      ? ` | Preset LoRA applied at generation: ${presetLoraDetails.map((lora) => lora.name ?? lora.id).join(", ")}`
-                      : " | No preset LoRAs"}
-                    {presetLoraDetails.some((lora) => lora.missing) ? " | Import still pending" : ""}
-                  </span>
-                </div>
-              ) : (
-                <div className="guidance-strip">
-                  <strong>No preset selected</strong>
-                  <span>Generation uses only the prompt, model, and visible render settings.</span>
-                </div>
-              )}
+              <PresetGuidanceStrip
+                selectedPreset={selectedPreset}
+                presetPromptParts={presetPromptParts}
+                presetLoraDetails={presetLoraDetails}
+                noPresetHint="Generation uses only the prompt, model, and visible render settings."
+              />
 
               {durationHint ? <p className="helper-copy">{durationHint}</p> : null}
